services/api/pyproject.toml
[project]
name = "TheResumeManager"
version = "0.1.0"
description = "A Personal Resume manger to Tackle with the troubles of managing the Resume for different job title for different employers."
authors = [
    {name = "Anshil P", email = "contactme@anshil.me"},
]
dependencies = [
    "qrcode>=7.4.2",
    "reportlab>=4.0.6",
    "PyPDF2>=3.0.1",
    "pdf2image>=1.16.3",
    "poppler-utils>=0.1.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.2",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "mysql-connector-python>=8.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
dev = "uvicorn src.theresumemanager.main:app --reload --host 0.0.0.0"
start = "gunicorn -k uvicorn.workers.UvicornWorker src.theresumemanager.main:app --host 0.0.0.0"